Refactored Day 3 to new format
