--- .github/workflows/dotnetcore_2019_day3.yml
@@ -1,29 +1,0 @@
-name: 2019 - Day 3
-
-on:
-  push:
-    branches: [ master ]
-    paths: "2019/Day3/**"
-  pull_request:
-    branches: [ master ]
-    paths: "2019/Day3/**"
-
-jobs:
-  Build_and_Test_2019_Day3:
-
-    runs-on: windows-latest
-
-    steps:
-    - uses: actions/checkout@v2
-    - name: Setup .NET Core
-      uses: actions/setup-dotnet@v1
-      with:
-        dotnet-version: 3.1.101
-    - name: Install dependencies for Solution
-      run: dotnet restore 2019/Day3/Solution/Day3.csproj
-    - name: Build
-      run: dotnet build --configuration Release --no-restore 2019/Day3/Solution/Day3.csproj
-    - name: Install dependencies for Tests
-      run: dotnet restore 2019/Day3/Tests/Day3Tests.csproj
-    - name: Test
-      run: dotnet test --no-restore --verbosity normal 2019/Day3/Tests/Day3Tests.csproj
