--- 2019/Day3/.vscode/extensions.json
@@ -1,8 +1,0 @@
-{
-	// List of extensions which should be recommended for users of this workspace.
-	"recommendations": [
-		"ms-dotnettools.csharp"
-	],
-	// List of extensions recommended by VS Code that should not be recommended for users of this workspace.
-	"unwantedRecommendations": []
-}
