--- 2019/Day3/Solution/Day3.csproj
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UserSecretsId>846e2ffd-4fb8-4da4-9405-d0f80b0ed87e</UserSecretsId>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="NAoCHelper" Version="1.2.5" />
-  </ItemGroup>
-</Project>
