--- 2019/Day3/Tests/Day3Tests.csproj
@@ -1,17 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <IsPackable>false</IsPackable>
-    <UserSecretsId>b25cce81-abd2-49c5-9e3e-703dc0aa6a40</UserSecretsId>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="nunit" Version="3.12.0" />
-    <PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.0" />
-    <PackageReference Include="Shouldly" Version="3.0.2" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\Solution\Day3.csproj" />
-  </ItemGroup>
-</Project>
