+++ 2019/Solutions/Day3/Dimensions.cs
@@ -1,0 +1,18 @@
+﻿namespace Solutions.Day3
+{
+    public class Dimensions
+    {
+        public int MinX { get; }
+        public int MinY { get; }
+        public int MaxX { get; }
+        public int MaxY { get; }
+
+        public Dimensions(int minX, int minY, int maxX, int maxY)
+        {
+            MinX = minX;
+            MinY = minY;
+            MaxX = maxX;
+            MaxY = maxY;
+        }
+    }
+}
